CORE-3066 chunk size to long, and add offset
We can assemble a binary blob on disk quicker if we have the offset
of each chunk, so that we can just write into place as we receive it.

Unlikely, but if file sizes are absurdly large, we ought to use long
rather than int for offsets and sizes, etc.