CORE-4292 Improved CPI uploading and error handling
Renamed UploadStatus to ChunkReceived for clarity.

Added ChunkKey because we need a unique identifier for each
ChunkReceived message.

Reduce enumeration to ok, failed, or in progress.

Add the part number back in.  We need this because we could have multiple
publishers and subscribers so the "compacted queue" of any subscriber needs
to be the complete message set.  An out-of-order receipt of previous
UploadStatus messages could lead to a perpetual "in progress" state.