CORE-5601 Unified external events
The sending on events to external workers is unified to simplify the handling of sending requests and receiving responses. This reduces the amount of code we have to write to add new functionality that requires communication with other workers.

As we are making a generic solution there will be limitations that each specific use-case requires, however a number of scenarios have already been identified and the solution _should_ cater for them all.

The generic code handles the sending events to other workers, receiving `ExternalEventResponse`s from them and determining whether to resume a flow or not based on the responses.

## Effects on the API

- `Checkpoint` now contains `ExternalEventState` rather than `CryptoState` and `PersistenceState`. This state will hold information about any type of external event that is to be sent from the flow worker.
- `ExternalEventContext` should be embedded into the events/payloads sent to external workers. This contains context information, which includes, a request id and flow id.
- `ExternalEventResponse`s are returned back to the flow worker after processing an external event. `ExternalEventResponse` contains an Avro serialized byte array of the underlying payload that the external processor returns to the flow worker.
- `ExternalEventResponseError` to convey error information back to the flow worker.