CORE-11065 Custom ledger queries liquibase changes
Add `custom` and `consumed_timestamp` to the
`utxo_relevant_transaction_state` table.

The `custom` column is of type `JSONB` and is therefore Postgres
specific.

`ledger-utxo-relevant-transaction-state-creation-v1.0.xml` has been
separated to create the table. For now it has been restricted to
creating the table only for Postgres databases, meaning that the table
will not be made for anything else. In the future, we can add the table
with the correct columns when we support more databases.