CORE-12912 Add `ResultSet.next` (#1072)
Add `ResultSet.next` and `ResultSet.hasNext`.

This allows users of the query api to iterate over all the pages of data that a query can return. This achieves two things, it simplifies the usage of the query api and fixes the issue where a vault query in-memory filters out all the results, making it look like there are no more pages to retrieve.