CORE-10329 Create new context map for non-member context properties (#1084)
This change introduces an additional context map to the internal network registration request. This is because the original context map is signed over and the signature becomes invalid if the map is modified so whatever is submitted in that context map became part of the member's member context map. It's preferred to not include the pre-auth token in the member info distributed across the network. To support this and other properties in the future which should not be distributed, an additional context map has been added.

This new context map is also signed. To avoid duplication, data and a signature over that data was pulled out in to it's own schema and membership schemas were updated where the same structure is already in use.