CORE-17360: Add topics for mapper scheduled cleanup and update cleanup Avro records (#1265)
Introduces new topics for scheduled cleanup in the flow mapper, and updates the execute cleanup record.

Two new topics are introduced. One is used for publishing scheduled task triggers to the flow mapper. The other is used to publish execute cleanup events to.

Additionally, the execute cleanup Avro record has been updated to allow multiple state IDs to be provided. This allows cleanups to be performed in batches if required.