CORE-17391 Add command metadata to registration state so command history can be tracked (#1269)
This change adds a command history to the registration state so that registration handlers can take appropriate action based on the previously completed stages. The main use case for this at the moment is to handle replays at the p2p level which cause repeat processing of registration commands. Using the previously processed command list, we can determine in a command processor whether a stage has already been processed successfully or not.