CORE-14025 Add new request to find existing transactions in database and their status (#1238)
Adding a new request to the UTXO ledger layer to fetch transaction alongside with their status.

This is needed when doing a back-chain resolution. We don't want to fetch transactions that are already present in our database. We just want to fetch their IDs though, no need to deserialise the whole object.