CORE-17958 Limit the length of client request ids to 242 characters. This string is combined with the vnode shorthash to make a key for flow mapper states records for start flow events. The flow mapper state is stored in the state storage which has a limit of 255 characters. vnode shorthash is 12 characters with an additional separator character added this adds to the max 255 (#1316)
This string is combined with the vnode shorthash to make a key for flow mapper states records for start flow events. The flow mapper state is stored in the state storage which has a limit of 255 characters. vnode shorthash is 12 characters with an additional separator character added this adds to the max 255.