CORE-18320 Support for a flow specific timeout on Flow Message API (#1394)
Support for a flow specific timeout on Flow Message API. Client can specify a flow session specific timeout on Flow Message API. If set, it will be passed to the initiated party to also use it. If the client does not provide a timeout value then the value from Corda Configuration is used.

A new class FlowSessionConfiguration is introduced to store session configuration. Methods that used configuration parameter requireClose will be deprecated with another PR.

Avro schema for SessionState was modified, field requireClose was removed as this value will be stored in sessionProperties.