CORE-19399 Update the membership registration state topic to be compacted (#1457)
The membership registration state topic for the state and event message pattern usage is the only state topic used in this pattern which is not compacted. This PR makes this topic compacted.
The impact of this not being compacted was that when a new consumer came online and all states were read into memory, every state ever written is read rather than just the latest state. This resulted in rebalances taking longer than necessary.