CORE-20186 Adding deduplication ID field to both persist APIs and update docs (#1574)
This PR changes the API of entity persistence in a breaking way. It's now required to pass a unique (within the context of a flow), deterministic deduplication ID with persist() (single or bulk) request to enable request idempotence in failure scenarios.

The maximum size of this new deduplication ID is 128 chars, and it will be maintained plaintext in the system (meaning that it will be visible as-is in the logs to aid debugging).